move diff command to archiver.diff
